requirements.txt

yfinance==0.2.28
pandas==2.1.4
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
tabulate==0.9.0
pytz==2023.3
pytz
pytz
